Reduced models to one file. Scaffolded out IdentityRepository.cs
