+++ Logistics/Satis.csproj
@@ -228,7 +228,6 @@
     <Compile Include="Helpers\MiscHelper.cs" />
     <Compile Include="Helpers\NotificationHelper.cs" />
     <Compile Include="Helpers\SessionHelper.cs" />
-    <Compile Include="Models\AccountViewModels.cs" />
     <Compile Include="Models\ErrorLogContext.cs" />
     <Compile Include="Models\AdminContext.cs" />
     <Compile Include="Controllers\AppointmentsController.cs" />
@@ -243,7 +242,6 @@
     <Compile Include="Areas\Config\Models\ConfigLookupContext.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\LocationContext.cs" />
-    <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Models\PersonnelContext.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\ErrorLogRepository.cs" />
